Restructure imports
